# src/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: API RESTful - Trisog
  version: 1.0.0
  description: >-
    This project involves the development of a **RESTful API using Node.js and
    TypeScript**, with a focus on integrating **MongoDB** for database
    management. The API is designed to manage **touristic experiences**,
    allowing users to interact with data related to destinations, activities,
    bookings, and reviews.


    The API will enable users to perform essential operations such as creating,
    retrieving, updating, and deleting information about destinations and
    activities. Additionally, it will support booking management and user
    reviews for these experiences. Security and data privacy are key priorities,
    with robust authentication and authorization mechanisms implemented using
    **Firebase Authentication**.


    **Firebase Storage** is utilized for managing and storing user-uploaded
    images associated with experiences. **Docker** and **Docker Compose** are
    employed to ensure consistent and scalable deployment across different
    environments. **Jest** is used for unit and integration testing, ensuring
    the reliability and quality of the API.


    **Prisma** is used as the ORM tool to facilitate interaction with MongoDB,
    providing a robust and type-safe interface for database queries and data
    manipulation. This allows for efficient and secure data handling throughout
    the application.


...


...


...
  license:
    name: MIT License
...
components:
  securitySchemes:
    jwt:
      type: apiKey
      in: header
      name: Authorization
      description: To authenticate, pass the JWT token in the format **Bearer _token_**.
  schemas:
...
    DeleteSuccess:
      type: object
      properties:
        msg:
          type: string
          description: Delete success message
          example: Successfully deleted
    BadRequest:
      type: object
      properties:
        msg:
          type: string
          description: Invalid request. Some parameters may be missing or invalid.
          example: Some parameters may be missing or invalid
    Unauthorized:
      type: object
      properties:
        msg:
          type: string
          description: Unauthorized. The JWT token was not provided or is invalid.
          example: JWT token not provided or invalid
    Forbidden:
      type: object
      properties:
        msg:
          type: string
          description: >-
            Forbidden. The user does not have permission to access the requested
            resource.
          example: You do not have permission to access the requested resource
    NotFound:
      type: object
      properties:
        msg:
          type: string
          description: No data found. The requested resource does not exist
          example: No data found
    InternalServerError:
      type: object
      properties:
        msg:
          type: string
          description: Internal error when handling data in the service
          example: Internal Server Error
    ReviewList:
      type: array
      items:
        $ref: '#/components/schemas/Review'
    Ratings:
      type: object
      required:
        - services
        - location
        - amenities
        - prices
        - food
        - room_comfort_and_quality
      properties:
        services:
          type: number
          description: Rating for services
          example: 4
        location:
          type: number
          description: Rating for location
          example: 5
        amenities:
          type: number
          description: Rating for amenities
          example: 3
        prices:
          type: number
          description: Rating for prices
          example: 4
        food:
          type: number
          description: Rating for food
          example: 5
        room_comfort_and_quality:
          type: number
          description: Rating for room comfort and quality
          example: 5
    Review:
      type: object
      required:
        - id
        - name
        - email
        - comment
        - image
        - ratings
        - created_at
        - experience_id
      properties:
        id:
          type: string
          description: Unique identifier for the review
          example: 60b8d295f1c0d2b0f1b2c3d8
        name:
          type: string
          description: Name of the person who wrote the review
          example: John Doe
        email:
          type: string
          description: Email address of the person who wrote the review
          example: johndoe@example.com
        comment:
          type: string
          description: The content of the review
          example: The tour was fantastic and exceeded my expectations!
        image:
          type: string
          description: URL to an image related to the review
          example: https://example.com/review-image.jpg
        ratings:
          $ref: '#/components/schemas/Ratings'
          description: Object containing individual ratings for various categories
        created_at:
          type: string
          format: date-time
          description: The date and time when the review was created
          example: '2024-08-27T14:30:00Z'
        experience_id:
          type: string
          description: Unique identifier for the experience being reviewed
          example: 60b8d295f1c0d2b0f1b2c3d9
    ReviewUpsert:
      type: object
      required:
        - name
        - email
        - comment
        - image
        - ratings
        - experience_id
      properties:
        name:
          type: string
          description: Name of the person who wrote the review
          example: John Doe
        email:
          type: string
          description: Email address of the person who wrote the review
          example: johndoe@example.com
        comment:
          type: string
          description: The content of the review
          example: The tour was fantastic and exceeded my expectations!
        image:
          type: string
          description: URL to an image related to the review
          example: https://example.com/review-image.jpg
        ratings:
          $ref: '#/components/schemas/Ratings'
          description: Object containing individual ratings for various categories
        experience_id:
          type: string
          description: Unique identifier for the experience being reviewed
          example: 60b8d295f1c0d2b0f1b2c3d9
...
paths:
...
  /reviews:
    get:
      summary: View all reviews
      description: Returns information about all reviews
      tags:
        - Review
      operationId: get_all_reviews
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewList'
        '500':
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    post:
      summary: Create a new review
      description: Creates a new review based on the data provided in the request body.
      tags:
        - Review
      operationId: create_review
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewUpsert'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '404':
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /reviews/{review_id}:
    parameters:
      - name: review_id
        in: path
        required: true
        description: ID of the review to be deleted
        schema:
          type: string
    get:
      summary: View review by id
      description: Returns review information by id
      tags:
        - Review
      operationId: get_review_by_id
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '404':
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    put:
      summary: Update review
      description: Updates the information of the desired review
      tags:
        - Review
      operationId: update_review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewUpsert'
      security:
        - jwt: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '404':
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
    delete:
      summary: Remove review
      description: Deletes a review based on the ID provided
      tags:
        - Review
      operationId: delete_review
      security:
        - jwt: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteSuccess'
        '400':
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '404':
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
  /reviews/experience/{experience_id}:
    parameters:
      - name: experience_id
        in: path
        required: true
        description: ID of the experience that will fetch the reviews
        schema:
          type: string
    get:
      summary: View review by experience
      description: Returns review information by experience
      tags:
        - Review
      operationId: get_review_by_experience
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewList'
        '400':
          description: BadRequest
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '404':
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: InternalServerError
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
...